+++ src/extension.ts
@@ -1,7 +1,8 @@
 import * as vscode from 'vscode'
-import { WorkspaceFolder, DebugConfiguration, ProviderResult, CancellationToken } from 'vscode'
+import { WorkspaceFolder, DebugConfiguration, CancellationToken } from 'vscode'
 import { LaunchRequestArguments } from './phpDebug'
 import * as which from 'which'
+import * as path from 'path'
 
 export function activate(context: vscode.ExtensionContext) {
     context.subscriptions.push(
@@ -10,15 +11,19 @@ export function activate(context: vscode.ExtensionContext) {
                 folder: WorkspaceFolder | undefined,
                 debugConfiguration: DebugConfiguration & LaunchRequestArguments,
                 token?: CancellationToken
-            ): Promise<ProviderResult<DebugConfiguration>> {
-                if (!debugConfiguration.type && !debugConfiguration.request && !debugConfiguration.name) {
+            ): Promise<DebugConfiguration | undefined> {
+                const isDynamic =
+                    (!debugConfiguration.type || debugConfiguration.type === 'php') &&
+                    !debugConfiguration.request &&
+                    !debugConfiguration.name
+                if (isDynamic) {
                     const editor = vscode.window.activeTextEditor
                     if (editor && editor.document.languageId === 'php') {
                         debugConfiguration.type = 'php'
                         debugConfiguration.name = 'Launch (dynamic)'
                         debugConfiguration.request = 'launch'
-                        debugConfiguration.program = '${file}'
-                        debugConfiguration.cwd = '${fileDirname}'
+                        debugConfiguration.program = debugConfiguration.program || '${file}'
+                        debugConfiguration.cwd = debugConfiguration.cwd || '${fileDirname}'
                         debugConfiguration.port = 0
                         debugConfiguration.runtimeArgs = ['-dxdebug.start_with_request=yes']
                         debugConfiguration.env = {
@@ -66,40 +71,71 @@ export function activate(context: vscode.ExtensionContext) {
                         }
                     }
                 }
+                if (folder && folder.uri.scheme !== 'file') {
+                    // replace
+                    if (debugConfiguration.pathMappings) {
+                        for (const key in debugConfiguration.pathMappings) {
+                            debugConfiguration.pathMappings[key] = debugConfiguration.pathMappings[key].replace(
+                                '${workspaceFolder}',
+                                folder.uri.toString()
+                            )
+                        }
+                    }
+                    // The following path are currently NOT mapped
+                    /*
+                    debugConfiguration.skipEntryPaths = debugConfiguration.skipEntryPaths?.map(v =>
+                        v.replace('${workspaceFolder}', folder.uri.toString())
+                    )
+                    debugConfiguration.skipFiles = debugConfiguration.skipFiles?.map(v =>
+                        v.replace('${workspaceFolder}', folder.uri.toString())
+                    )
+                    debugConfiguration.ignore = debugConfiguration.ignore?.map(v =>
+                        v.replace('${workspaceFolder}', folder.uri.toString())
+                    )
+                    */
+                }
                 return debugConfiguration
             },
         })
     )
+
     context.subscriptions.push(
-        vscode.languages.registerEvaluatableExpressionProvider('php', {
-            // eslint-disable-next-line @typescript-eslint/require-await
-            async provideEvaluatableExpression(
-                document: vscode.TextDocument,
-                position: vscode.Position,
-                token: CancellationToken
-            ): Promise<ProviderResult<vscode.EvaluatableExpression>> {
-                // see https://www.php.net/manual/en/language.variables.basics.php
-                // const wordRange = document.getWordRangeAtPosition(position, /\$([a-zA-Z_\x80-\xff][a-zA-Z0-9_\x80-\xff]*)((->(?1))|\[(\d+|'[^']+'|"[^"]+"|(?0))\])*/)
-                const wordRange = document.getWordRangeAtPosition(
-                    position,
-                    /\$[a-zA-Z_\x80-\xff][a-zA-Z0-9_\x80-\xff]*(->[a-zA-Z_\x80-\xff][a-zA-Z0-9_\x80-\xff]*)*/
-                )
-                if (wordRange) {
-                    return new vscode.EvaluatableExpression(wordRange)
-                }
-                return undefined // nothing evaluatable found under mouse
-            },
+        vscode.commands.registerCommand('extension.php-debug.runEditorContents', (resource: vscode.Uri) => {
+            let targetResource = resource
+            if (!targetResource && vscode.window.activeTextEditor) {
+                targetResource = vscode.window.activeTextEditor.document.uri
+            }
+            if (targetResource) {
+                void vscode.debug.startDebugging(undefined, {
+                    type: 'php',
+                    name: '',
+                    request: '',
+                    noDebug: true,
+                    program: targetResource.fsPath,
+                    cwd: path.dirname(targetResource.fsPath),
+                })
+            }
+        }),
+        vscode.commands.registerCommand('extension.php-debug.debugEditorContents', (resource: vscode.Uri) => {
+            let targetResource = resource
+            if (!targetResource && vscode.window.activeTextEditor) {
+                targetResource = vscode.window.activeTextEditor.document.uri
+            }
+            if (targetResource) {
+                void vscode.debug.startDebugging(undefined, {
+                    type: 'php',
+                    name: '',
+                    request: '',
+                    stopOnEntry: true,
+                    program: targetResource.fsPath,
+                    cwd: path.dirname(targetResource.fsPath),
+                })
+            }
         })
     )
 
     context.subscriptions.push(
-        vscode.commands.registerCommand('php.debug.debugPhpFile', async (uri: vscode.Uri) => {
-            await vscode.debug.startDebugging(undefined, { type: '', name: '', request: '' })
-        })
-    )
-
-    context.subscriptions.push(
-        vscode.commands.registerCommand('php.debug.startWithStopOnEntry', async (uri: vscode.Uri) => {
+        vscode.commands.registerCommand('extension.php-debug.startWithStopOnEntry', async (uri: vscode.Uri) => {
             await vscode.commands.executeCommand('workbench.action.debug.start', {
                 config: {
                     stopOnEntry: true,