+++ src/phpDebug.ts
@@ -79,6 +79,8 @@ export interface LaunchRequestArguments extends VSCodeDebugProtocol.LaunchReques
     ignore?: string[]
     /** Array of glob patterns that exceptions should be ignored from */
     ignoreExceptions?: string[]
+    /** An array of glob pattern to skip if the initial entry file is matched. */
+    skipEntryPaths?: string[]
     /** Array of glob patterns that debugger should not step in */
     skipFiles?: string[]
     /** Xdebug configuration */
@@ -96,6 +98,10 @@ export interface LaunchRequestArguments extends VSCodeDebugProtocol.LaunchReques
     maxConnections?: number
     /** Xdebug cloud token */
     xdebugCloudToken?: string
+    /** Xdebug stream settings */
+    stream?: {
+        stdout?: 0 | 1 | 2
+    }
 
     // CLI options
 
@@ -413,6 +419,11 @@ class PhpDebugSession extends vscode.DebugSession {
                     throw new Error('Cannot have port and socketPath set at the same time')
                 }
                 if (args.hostname?.toLowerCase()?.startsWith('unix://') === true) {
+                    if (fs.existsSync(args.hostname.substring(7))) {
+                        throw new Error(
+                            `File ${args.hostname.substring(7)} exists and cannot be used for Unix Domain socket`
+                        )
+                    }
                     server.listen(args.hostname.substring(7))
                 } else if (args.hostname?.startsWith('\\\\') === true) {
                     server.listen(args.hostname)
@@ -488,6 +499,26 @@ class PhpDebugSession extends vscode.DebugSession {
     private async initializeConnection(connection: xdebug.Connection): Promise<void> {
         const initPacket = await connection.waitForInitPacket()
 
+        // check if this connection should be skipped
+        if (
+            this._args.skipEntryPaths &&
+            isPositiveMatchInGlobs(
+                convertDebuggerPathToClient(initPacket.fileUri).replace(/\\/g, '/'),
+                this._args.skipEntryPaths
+            )
+        ) {
+            this.sendEvent(
+                new vscode.OutputEvent(
+                    `skipping entry point ${convertDebuggerPathToClient(initPacket.fileUri).replace(
+                        /\\/g,
+                        '/'
+                    )} on connection ${connection.id}\n`
+                )
+            )
+            this.disposeConnection(connection)
+            return
+        }
+
         // support for breakpoints
         let feat: xdebug.FeatureGetResponse
         const supportedEngine =
@@ -538,6 +569,15 @@ class PhpDebugSession extends vscode.DebugSession {
             this._args.xdebugSettings = xdebugSettings
         } catch (error) {
             throw new Error(`Error applying xdebugSettings: ${String(error instanceof Error ? error.message : error)}`)
+        }
+
+        const stdout =
+            this._args.stream?.stdout === undefined ? (this._args.externalConsole ? 1 : 0) : this._args.stream.stdout
+        if (stdout) {
+            await connection.sendStdout(stdout)
+            connection.on('stream', (stream: xdebug.Stream) =>
+                this.sendEvent(new vscode.OutputEvent(stream.value, 'stdout'))
+            )
         }
 
         this.sendEvent(new vscode.ThreadEvent('started', connection.id))