+++ CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.35.0]
+
+- Support for DBGp stream command
+- Avoid conflict with full screen F11 shortcut
+- Improve existing unix socket handling
+- Improve close socket handling
+
+## [1.34.0]
+
+- Partial support for virtual workspaces
+
+## [1.33.1]
+
+- Fix editor title run/debug button.
+
+## [1.33.0]
+
+- Add skipEntryPaths to immediately detach a debug session depending on entry path.
+- Remove EvaluatableExpression Provider.
+
 ## [1.32.1]
 
 - Fix logging of cloud connection.